[Feature] Spring Security, JWT 인증, 인가 (#25)
* update: SignupRequestDto Validation 어노테이션 추가

* update: SignupRequestDto nickname 최대 50자 조건 추가

* build: JWT 의존성 추가

* update: Spring Security 기능 다시 켜기

* feat: Spring Security 설정 및 Login API 구현

* feat: Todo 등록시 Member 관계 연결

* feat: Todo 조회 시 Member 관계 연결

* feat: Todo 수정 시 Member 관계 연결

* feat: Todo 삭제 시 Member 관계 연결

+++ src/main/java/sesac/semiProject/todo/service/TodoServiceImpl.java
@@ -3,7 +3,6 @@ package sesac.semiProject.todo.service;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
 
 import jakarta.transaction.Transactional;
@@ -12,6 +11,7 @@ import sesac.semiProject.common.constants.TodoConstants;
 import sesac.semiProject.common.dto.ApiResponseDto;
 import sesac.semiProject.common.exception.CustomErrorCode;
 import sesac.semiProject.common.exception.CustomException;
+import sesac.semiProject.member.model.Member;
 import sesac.semiProject.todo.dto.TodoRequestDto;
 import sesac.semiProject.todo.dto.TodoResponseDto;
 import sesac.semiProject.todo.model.Todo;
@@ -24,28 +24,34 @@ public class TodoServiceImpl implements TodoService {
     private final TodoRepository todoRepository;
 
     @Override
-    public TodoResponseDto createTodo(TodoRequestDto requestDto) {
-        Todo todo = requestDto.toEntity();
+    public TodoResponseDto createTodo(TodoRequestDto requestDto, Member member) {
+        Todo todo = requestDto.toEntity(member);
         Todo savedTodo = todoRepository.save(todo);
         return savedTodo.toDto();
     }
 
     @Override
-    public List<TodoResponseDto> getAllTodos() {
-        List<Todo> todos = todoRepository.findAllByOrderByIdAsc();
+    public List<TodoResponseDto> getAllTodos(Member member) {
+        List<Todo> todos = todoRepository.findAllByMember_IdOrderByIdAsc(member.getId());
         return todos.stream().map(Todo::toDto).toList();
     }
 
     @Override
-    public TodoResponseDto updateTodo(int id, TodoRequestDto requestDto) {
+    public TodoResponseDto updateTodo(int id, TodoRequestDto requestDto, Member member) {
         Todo todo = findTodo(id);
+        if(!todo.getMember().getId().equals(member.getId())) {
+            throw new CustomException(CustomErrorCode.NOT_AUTHORIZED);
+        }
         todo.update(requestDto);
         return todo.toDto();
     }
 
     @Override
-    public ApiResponseDto deleteTodo(int id) {
+    public ApiResponseDto deleteTodo(int id, Member member) {
         Todo todo = findTodo(id);
+        if(!todo.getMember().getId().equals(member.getId())) {
+            throw new CustomException(CustomErrorCode.NOT_AUTHORIZED);
+        }
         todoRepository.deleteById(id);
         return new ApiResponseDto(HttpStatus.OK.value(), TodoConstants.DELETE_TODO_SUCCESS);
     }
